+++ mldaikon/trace/types.py
@@ -11,7 +11,7 @@ class VarInstId(NamedTuple):
 
 
 class Liveness:
-    def __init__(self, start_time: int | None, end_time: int | None):
+    def __init__(self, start_time: float | None, end_time: float | None):
         self.start_time = start_time
         self.end_time = end_time
 
@@ -90,7 +90,7 @@ class FuncCallEvent(HighLevelEvent):
 class IncompleteFuncCallEvent(HighLevelEvent):
     """An outermost function call event, but without the post record."""
 
-    def __init__(self, func_name: str, pre_record: dict, potential_end_time: int):
+    def __init__(self, func_name: str, pre_record: dict, potential_end_time: float):
         self.func_name = func_name
         self.pre_record = pre_record
         self.potential_end_time = potential_end_time
@@ -138,7 +138,7 @@ class VarChangeEvent(HighLevelEvent):
         self,
         var_id: VarInstId,
         attr_name: str,
-        change_time: int,
+        change_time: float,
         old_state: AttrState,
         new_state: AttrState,
     ):
